move PrivacyIcon to global components folder
