+++ projects/packages/videopress/src/client/admin/components/video-row/index.tsx
@@ -5,8 +5,8 @@ import { sprintf, __ } from '@wordpress/i18n';
 import { Icon, image, trash, chevronDown, chevronUp } from '@wordpress/icons';
 import classNames from 'classnames';
 import { useState, useRef } from 'react';
+import privacy from '../../../components/icons/privacy-icon';
 import Checkbox from '../checkbox';
-import privacy from './privacy-icon';
 import StatsBase from './stats';
 import styles from './style.module.scss';
 import { VideoPressVideo, VideoRowProps } from './types';
