+++ projects/packages/videopress/src/client/components/icons/privacy-icon.js
